% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addHGNC.R
\name{addHGNC}
\alias{addHGNC}
\title{addHGNC.R}
\usage{
addHGNC(myDB, martDF)
}
\arguments{
\item{myDB}{A list. The database which is being built.}

\item{martDF}{The dataframe containing BiomaRt data.}
}
\value{
myDB with the HGNC data added to the database.
}
\description{
\code{addHGNC} Add the HGNC symbols to the data set along
with the corresponding gene description, stable ID,
stable version ID and gene name.
}
\details{
Details.
}
