% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPDB.R
\name{addPDB}
\alias{addPDB}
\title{addPDB.R}
\usage{
addPDB(myDB, martDF)
}
\arguments{
\item{myDB}{A list. The database which is being built.}

\item{martDF}{The dataframe containing BiomaRt data.}
}
\value{
myDB with the PDB IDs added to the database.
}
\description{
\code{addPDB} Add the PDB IDs and the corresponding PDB chains in that
PDB ID to the database.
}
\details{
Details.
}
