% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addpdbChain.R
\name{addpdbChain}
\alias{addpdbChain}
\title{addpdbChain.R}
\usage{
addpdbChain(myDB, martDF)
}
\arguments{
\item{myDB}{A list. The database which is being built.}

\item{martDF}{The dataframe containing BiomaRt data.}
}
\value{
myDB with the pdbChains added to the database.
}
\description{
\code{addpdbChain} Add the PDB chains and the corresponding HGNC ID
of the genes and transcripts the PDB chain corresponds to. As well,
add the PDB IDs of the PDB chains.
}
\details{
Details.
}
