% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPDBdata.R
\name{addPDBdata}
\alias{addPDBdata}
\title{addPDBdata.R}
\usage{
addPDBdata(myDB)
}
\arguments{
\item{myDB}{A list. The database which is being built.}
}
\value{
myDB with the PDB data added to the database.
}
\description{
\code{addPDBdata} Add the PDB data retrieved from PDB. This include
The sequence and Resolution of each PDB chain in the database.
}
\details{
Details.
}
