% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTranscripts.R
\name{addTranscripts}
\alias{addTranscripts}
\title{addTranscripts.R}
\usage{
addTranscripts(myDB, martDF)
}
\arguments{
\item{myDB}{A list. The database which is being built.}

\item{martDF}{The dataframe containing BiomaRt data.}
}
\value{
myDB with the Transcript data added to the database.
}
\description{
\code{addTranscripts} Add the transcript data to the database.
This includes the HGNC ID for the transcript, the stable ID, the
stable ID with the version number, The corresponding HGNC symbols,
and the start and end gene coordinates.
}
\details{
Details.
}
