% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{getData.R}
\usage{
getData()
}
\value{
martDF; a dataframe containing the biomaRt data
for homosapien genes
}
\description{
\code{getData} Retrieve the data from the mart_export.txt file
from BiomaRts webpage and download services. Guidelines
provided in the readMe. \code{getData} returns a
dataframe containing the biomaRt data.
}
\details{
Details.
}
