% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDBdataset.R
\name{PDBdataset}
\alias{PDBdataset}
\title{PDBdataset.R}
\usage{
PDBdataset()
}
\description{
\code{PDBdataset} Generates a PDB-HGNC mapping database linking HGNC transcripts
to PDB transcript identifiers. The resulting database also contains several
annotations about each transcript such as it's basic description and
PDB sequences.
}
