% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readXML.R
\name{readXML}
\alias{readXML}
\title{readXML.R}
\usage{
readXML(URL)
}
\arguments{
\item{URL}{The URL to the XML data from the PDB webservices in the
form of a character vector}
}
\value{
The dataframe containing the protein structure ID and
the corresponding resolution
}
\description{
Retrieve the XML data from the PDB web services
and generate a dataframe containing the resolution
and StructureID for a PDB structure of interest
}
