% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{validate.R}
\usage{
validate(transcriptHGNC, chainID, myDB, gffData, showConsensus = TRUE)
}
\arguments{
\item{transcriptHGNC}{The stable transcript ID of a transcript}

\item{chainID}{The ID of the PDB chain}

\item{myDB}{The annotated database loaded in the environment}

\item{gffData}{A dataframe containing the Gencode data imported
manually by the user}

\item{showConsensus}{A boolean value determining if the user
would like to see the consensus sequence.
Set to TRUE automatically}
}
\description{
This function provides users with validation on the
PDB and HGNC database generated from the BCB420.2019.PDB
package. It performs a MSA between the provided Transcript
and PDB chain sequences to determine if the two sequenes are
identical. It also verifies that for both the
Transcript and PDB chain map to the same gene coordinates.
}
